Add scroll navigation to main page and update Navbar
Added a scroll navigation feature to the main page, including dots on the side that follow the scroll position. The navigation bar ("Navbar") has also been updated with new links and a custom font. Several other cosmetic changes were implemented as well, and the 'jquery' and 'react-scroll' dependencies were added to the package.json file.